Rewrite dnf-testing script from shell to python
Main new feature is that destructive tests are run one scenario per container
The output stays as compatible to the shell version as possible, because
the script is used also in Jenkins test job and in beaker tests.
The current workflows should not be broken as exists thin wrapper
dnf-testing.sh around new python implemantation.